requirements.txt
discord.py==2.3.2
firebase-admin==6.5.0
matplotlib==3.9.0
